requirements.txt
aiogram
python-dotenv
torch
openai-whisper
g4f
pydub
PyMuPDF
pypdf
python-docx